+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- LocalizationManager.Select using string identifier
+
 ## [1.3.0] - 2025-10-10
 ### Added
 - GameDataLocaleHandler